+++ app/interactors/broadcaster/create_subscription.rb
@@ -1,0 +1,14 @@
+module Broadcaster
+  class CreateSubscription
+    include Interactor
+
+    def call
+      Broadcaster::Subscription.create!(
+        subscriber: subscriber,
+        broadcaster_campaign: broadcaster_campaign,
+      )
+    end
+
+    delegate :subscriber, :broadcaster_campaign, to: :context
+  end
+end