+++ db/migrate/20190619075929_add_index_to_subscriptions.rb
@@ -3,6 +3,6 @@ class AddIndexToSubscriptions < ActiveRecord::Migration
     add_index :broadcaster_subscriptions,
               %i[subscriber_id subscriber_type broadcaster_campaign_id],
               unique: true,
-              name: :index_broadcaster_subscriptions_on_subscriber_campaign
+              name: :index_broadcaster_subscriptions_on_subscriber_and_campaign
   end
 end