+++ spec/dummy/db/schema.rb
@@ -21,7 +21,7 @@ ActiveRecord::Schema.define(version: 20190619075938) do
     t.string   "label"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.index ["publisher_id", "publisher_type", "label"], name: "index_broadcaster_campaign_on_publisher_and_label", unique: true, using: :btree
+    t.index ["publisher_id", "publisher_type", "label"], name: "index_broadcaster_campaigns_on_publisher_and_label", unique: true, using: :btree
     t.index ["publisher_type", "publisher_id"], name: "index_broadcaster_campaigns_on_publisher_type_and_publisher_id", using: :btree
   end
 
@@ -30,11 +30,11 @@ ActiveRecord::Schema.define(version: 20190619075938) do
     t.integer  "subscriber_id"
     t.integer  "broadcaster_campaign_id"
     t.string   "unsubscribe_token"
-    t.boolean  "active"
+    t.boolean  "active",                  default: true
     t.datetime "created_at"
     t.datetime "updated_at"
     t.index ["broadcaster_campaign_id"], name: "index_broadcaster_subscriptions_on_broadcaster_campaign_id", using: :btree
-    t.index ["subscriber_id", "subscriber_type", "broadcaster_campaign_id"], name: "index_broadcaster_subscriptions_on_subscriber_campaign", unique: true, using: :btree
+    t.index ["subscriber_id", "subscriber_type", "broadcaster_campaign_id"], name: "index_broadcaster_subscriptions_on_subscriber_and_campaign", unique: true, using: :btree
     t.index ["subscriber_type", "subscriber_id"], name: "index_subscriptions_on_subscriber", using: :btree
   end